requirements.txt
# Core dependencies - Python 3.13.9 Compatible with resolved conflicts
click>=8.1.0,<8.2.2  # Compatible with bump-my-version requirement
colorama>=0.4.6
tomli>=2.0.1
tomli-w>=1.0.0
tqdm>=4.66.0

# Image processing and computer vision - Python 3.13.9 Optimized
opencv-python>=4.9.0  # Full Python 3.13 support
Pillow>=10.2.0  # Latest stable for Python 3.13
scikit-image>=0.22.0  # Python 3.13 compatible
piexif>=1.1.3

# Data processing and analysis - Python 3.13.9 Compatible with nptyping constraint
numpy>=1.20.0,<2.0.0  # Compatible with nptyping requirement and Python 3.13
pandas>=2.0.3  # Python 3.13 compatible
scipy>=1.11.0  # Stable for Python 3.13
scikit-learn>=1.3.0  # Stable for Python 3.13
matplotlib>=3.7.2  # Stable for Python 3.13

# AI and Machine Learning - Python 3.13.9 Ready
torch>=2.1.0  # Stable PyTorch with Python 3.13 support
transformers>=4.30.0  # Stable with Python 3.13 compatibility
ultralytics>=8.0.100  # Stable YOLO version

# Database and persistence
sqlalchemy>=2.0.19  # Stable for Python 3.13

# Parallel processing
joblib>=1.3.1  # Stable for Python 3.13

# Document generation
reportlab>=4.0.4  # Stable for Python 3.13

# GUI (PyQt6) - Python 3.13.9 Fully Supported
PyQt6>=6.5.2  # Stable Python 3.13 support
PyQt6-Qt6>=6.5.2
PyQt6-sip>=13.5.2

# Web interface (optional)
fastapi>=0.101.0  # Stable for Python 3.13
uvicorn>=0.23.0  # Stable for Python 3.13

# Language tools
language-tool-python>=2.7.1
pyspellchecker>=0.7.2

# Windows-specific
pywin32>=306; sys_platform=="win32"
pywin32>=306; sys_platform=="win64"

# Development and testing
pytest>=7.4.0  # Stable for Python 3.13
pytest-cov>=4.1.0

# EXIF data handling
exifread>=3.0.0

# Build tools - Python 3.13.9 Compatible with pths constraint
setuptools>=68.2.2,<69.0.0  # Compatible with pths requirement
wheel>=0.40.0  # Stable wheel version
packaging>=23.1  # Stable packaging
